feat(sidebar): provide a proxy for the files sidebar
Additionally to registering sidebar tabs this also provides
a public interface to access the sidebar itself.
Which can be used to retrieve or set the currently active tab,
allows to open or close it and getting the open state.

We can then use this proxy to later add fallbacks for different
Nextcloud versions once we modernized the files sidebar,
so apps can stick with this interface regardless of the Nextcloud
version and implementation used underneath.
